ltoplugin: split pass registration
