+++ ltoplugin/LLGOLTOPlugin.cpp
@@ -1,0 +1,31 @@
+#include "LLGOLTOPasses.h"
+
+#include "llvm/Config/llvm-config.h"
+#include "llvm/Passes/PassBuilder.h"
+#include "llvm/Passes/PassPlugin.h"
+#include "llvm/Support/Compiler.h"
+
+using namespace llvm;
+
+PassPluginLibraryInfo getLLGOLTOPluginInfo() {
+  return {LLVM_PLUGIN_API_VERSION, "llgo-lto-plugin", LLVM_VERSION_STRING,
+          [](PassBuilder &PB) {
+            PB.registerPipelineParsingCallback(
+                [](StringRef Name, ModulePassManager &MPM,
+                   ArrayRef<PassBuilder::PipelineElement>) {
+                  if (Name != llgo::LLGOPreGlobalDCEPassName)
+                    return false;
+                  llgo::addLLGOPreGlobalDCEPipeline(MPM);
+                  return true;
+                });
+
+            PB.registerFullLinkTimeOptimizationEarlyEPCallback(
+                [](ModulePassManager &MPM, OptimizationLevel) {
+                  llgo::addLLGOPreGlobalDCEPipeline(MPM);
+                });
+          }};
+}
+
+extern "C" LLVM_ATTRIBUTE_WEAK PassPluginLibraryInfo llvmGetPassPluginInfo() {
+  return getLLGOLTOPluginInfo();
+}
